extract atoms-specific e2e tests into their own suite
Split the atom-family tests (locators, element state, interaction,
storage, script execution) out of safari-e2e.spec.ts into a new
atoms-e2e.spec.ts, sharing simulator/RemoteDebugger setup via a new
rd-fixture.ts helper. Add a dedicated e2e-test:atoms npm script and
point verify-atoms.yml's e2e job at it, so atoms changes only run the
atoms-specific suite instead of the full functional-test.yml suite.
functional-test.yml keeps running the remaining safari-e2e.spec.ts
tests (connection handling, network/console capture, screenshots,
shadow DOM) unconditionally, as before.
